+++ xarray/tests/test_backends.py
@@ -2186,6 +2186,16 @@ class ZarrBase(CFEncodedBase):
                 with self.open(store) as actual:
                     assert_identical(xr.concat([ds, ds_to_append], dim="time"), actual)
 
+    @pytest.mark.parametrize("chunk", [False, True])
+    def test_save_emptydim(self, chunk):
+        if chunk and not has_dask:
+            pytest.skip("requires dask")
+        ds = Dataset({"x": (("a", "b"), np.empty((5, 0))), "y": ("a", [1, 2, 5, 8, 9])})
+        if chunk:
+            ds = ds.chunk({})  # chunk dataset to save dask array
+        with self.roundtrip(ds) as ds_reload:
+            assert_identical(ds, ds_reload)
+
     @pytest.mark.parametrize("consolidated", [False, True])
     @pytest.mark.parametrize("compute", [False, True])
     @pytest.mark.parametrize("use_dask", [False, True])